+++ Tests/AnchorDimensionTests.swift
@@ -17,33 +17,33 @@ class AnchorDimensionTests: XCTestCase {
 
     func testSetWidth() {
         XCTAssertEqualConstraints(
-            view.al.width.set(10),
+            view.anchors.width.set(10),
             NSLayoutConstraint(item: view, attribute: .width, constant: 10)
         )
         XCTAssertEqualConstraints(
-            view.al.width.set(10, relation: .greaterThanOrEqual),
+            view.anchors.width.set(10, relation: .greaterThanOrEqual),
             NSLayoutConstraint(item: view, attribute: .width, relation: .greaterThanOrEqual, constant: 10)
         )
     }
 
     func testSetHeight() {
         XCTAssertEqualConstraints(
-            view.al.height.set(10),
+            view.anchors.height.set(10),
             NSLayoutConstraint(item: view, attribute: .height, constant: 10)
         )
         XCTAssertEqualConstraints(
-            view.al.height.set(10, relation: .greaterThanOrEqual),
+            view.anchors.height.set(10, relation: .greaterThanOrEqual),
             NSLayoutConstraint(item: view, attribute: .height, relation: .greaterThanOrEqual, constant: 10)
         )
     }
 
     func testMatchWidth() {
         XCTAssertEqualConstraints(
-            view.al.width.match(container.al.width),
+            view.anchors.width.match(container.anchors.width),
             NSLayoutConstraint(item: view, attribute: .width, toItem: container, attribute: .width)
         )
         XCTAssertEqualConstraints(
-            view.al.width.match(container.al.height), // can mix and match
+            view.anchors.width.match(container.anchors.height), // can mix and match
             NSLayoutConstraint(item: view, attribute: .width, toItem: container, attribute: .height)
         )
     }