+++ Tests/AnchorEdgeTests.swift
@@ -18,41 +18,41 @@ class AnchorEdgeTests: XCTestCase {
 
     func testPinToSuperview() {
         test("pin top to superview") {
-            let c = view.al.top.pinToSuperview()
+            let c = view.anchors.top.pinToSuperview()
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .top, toItem: container, attribute: .top))
         }
 
         test("pin top to superview with inset") {
-            let c = view.al.top.pinToSuperview(inset: 10)
+            let c = view.anchors.top.pinToSuperview(inset: 10)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .top, toItem: container, attribute: .top, constant: 10))
         }
         test("pin bottom to superview with inset") {
-            let c = view.al.bottom.pinToSuperview(inset: 10)
+            let c = view.anchors.bottom.pinToSuperview(inset: 10)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .bottom, toItem: container, attribute: .bottom, constant: -10))
         }
         test("pin left to superview with inset") {
-            let c = view.al.left.pinToSuperview(inset: 10)
+            let c = view.anchors.left.pinToSuperview(inset: 10)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .left, toItem: container, attribute: .left, constant: 10))
         }
         test("pin right to superview with inset") {
-            let c = view.al.right.pinToSuperview(inset: 10)
+            let c = view.anchors.right.pinToSuperview(inset: 10)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .right, toItem: container, attribute: .right, constant: -10))
         }
 
         test("pin top to superview with inset and relation") {
-            let c = view.al.top.pinToSuperview(inset: 10, relation: .greaterThanOrEqual)
+            let c = view.anchors.top.pinToSuperview(inset: 10, relation: .greaterThanOrEqual)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .top, relation: .greaterThanOrEqual, toItem: container, attribute: .top, constant: 10))
         }
         test("pin bottom to superview with inset and relation") {
-            let c = view.al.bottom.pinToSuperview(inset: 10, relation: .greaterThanOrEqual)
+            let c = view.anchors.bottom.pinToSuperview(inset: 10, relation: .greaterThanOrEqual)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .bottom, relation: .lessThanOrEqual, toItem: container, attribute: .bottom, constant: -10))
         }
         test("pin left to superview with inset and relation") {
-            let c = view.al.left.pinToSuperview(inset: 10, relation: .greaterThanOrEqual)
+            let c = view.anchors.left.pinToSuperview(inset: 10, relation: .greaterThanOrEqual)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .left, relation: .greaterThanOrEqual, toItem: container, attribute: .left, constant: 10))
         }
         test("pin right to superview with inset and relation") {
-            let c = view.al.right.pinToSuperview(inset: 10, relation: .greaterThanOrEqual)
+            let c = view.anchors.right.pinToSuperview(inset: 10, relation: .greaterThanOrEqual)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .right, relation: .lessThanOrEqual, toItem: container, attribute: .right, constant: -10))
         }
     }
@@ -60,41 +60,41 @@ class AnchorEdgeTests: XCTestCase {
 #if os(iOS) || os(tvOS)
     func testPinToSuperviewMargin() {
         test("pin top to superview margin") {
-            let c = view.al.top.pinToSuperviewMargin()
+            let c = view.anchors.top.pinToSuperviewMargin()
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .top, toItem: container, attribute: .topMargin))
         }
 
         test("pin top to superview margin with inset") {
-            let c = view.al.top.pinToSuperviewMargin(inset: 10)
+            let c = view.anchors.top.pinToSuperviewMargin(inset: 10)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .top, toItem: container, attribute: .topMargin, constant: 10))
         }
         test("pin bottom to superview margin with inset") {
-            let c = view.al.bottom.pinToSuperviewMargin(inset: 10)
+            let c = view.anchors.bottom.pinToSuperviewMargin(inset: 10)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .bottom, toItem: container, attribute: .bottomMargin, constant: -10))
         }
         test("pin left to superview margin with inset") {
-            let c = view.al.left.pinToSuperviewMargin(inset: 10)
+            let c = view.anchors.left.pinToSuperviewMargin(inset: 10)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .left, toItem: container, attribute: .leftMargin, constant: 10))
         }
         test("pin right to superview margin with inset") {
-            let c = view.al.right.pinToSuperviewMargin(inset: 10)
+            let c = view.anchors.right.pinToSuperviewMargin(inset: 10)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .right, toItem: container, attribute: .rightMargin, constant: -10))
         }
 
         test("pin top to superview margin with inset and relation") {
-            let c = view.al.top.pinToSuperviewMargin(inset: 10, relation: .greaterThanOrEqual)
+            let c = view.anchors.top.pinToSuperviewMargin(inset: 10, relation: .greaterThanOrEqual)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .top, relation: .greaterThanOrEqual, toItem: container, attribute: .topMargin, constant: 10))
         }
         test("pin bottom to superview margin with inset and relation") {
-            let c = view.al.bottom.pinToSuperviewMargin(inset: 10, relation: .greaterThanOrEqual)
+            let c = view.anchors.bottom.pinToSuperviewMargin(inset: 10, relation: .greaterThanOrEqual)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .bottom, relation: .lessThanOrEqual, toItem: container, attribute: .bottomMargin, constant: -10))
         }
         test("pin left to superview margin with inset and relation") {
-            let c = view.al.left.pinToSuperviewMargin(inset: 10, relation: .greaterThanOrEqual)
+            let c = view.anchors.left.pinToSuperviewMargin(inset: 10, relation: .greaterThanOrEqual)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .left, relation: .greaterThanOrEqual, toItem: container, attribute: .leftMargin, constant: 10))
         }
         test("pin right to superview margin with inset and relation") {
-            let c = view.al.right.pinToSuperviewMargin(inset: 10, relation: .greaterThanOrEqual)
+            let c = view.anchors.right.pinToSuperviewMargin(inset: 10, relation: .greaterThanOrEqual)
             XCTAssertEqualConstraints(c, NSLayoutConstraint(item: view, attribute: .right, relation: .lessThanOrEqual, toItem: container, attribute: .rightMargin, constant: -10))
         }
     }